+++ src/ir/module-utils.cpp
@@ -54,7 +54,7 @@ Function* copyFunction(Function* func,
   ret->localNames = func->localNames;
   ret->localIndices = func->localIndices;
   ret->body = ExpressionManipulator::copy(func->body, out);
-  debug::copyDebugInfo(func->body, ret->body, func, ret.get());
+  debug::copyDebugInfoBetweenFunctions(func->body, ret->body, func, ret.get());
   ret->prologLocation = func->prologLocation;
   ret->epilogLocation = func->epilogLocation;
   // Update file indices if needed
