+++ src/passes/Inlining.cpp
@@ -579,7 +579,7 @@ static Expression* doInlining(Module* module,
 
   // Generate and update the inlined contents
   auto* contents = ExpressionManipulator::copy(from->body, *module);
-  debug::copyDebugInfo(from->body, contents, from, into);
+  debug::copyDebugInfoBetweenFunctions(from->body, contents, from, into);
   updater.walk(contents);
   block->list.push_back(contents);
   block->type = retType;
